Add initial S3 backend and refactor structure.
Create an initial backend interface and S3 implementation. Move the
Bucket model into the common folder with the backends because that's
where it'll be used most. Update the bucket component imports.
