+++ src/common/AmazonS3Backend.ts
@@ -1,0 +1,16 @@
+import {Bucket} from './Bucket';
+import {IBackend} from './IBackend';
+
+export class AmazonS3Backend extends IBackend {
+  /**
+   * Create an instance of the backend.
+   */
+  constructor(public region: string) { }
+
+  /**
+   * Get a bucket with the given name.
+   */
+  getBucket(name: string): Bucket {
+    return new Bucket(name);
+  }
+}
