+++ src/common/IBackend.ts
@@ -1,0 +1,8 @@
+import {Bucket} from './Bucket';
+
+export interface IBackend {
+  /**
+   * Get a bucket with the given name.
+   */
+  getBucket(name: string): Bucket;
+}
